// copy-of-木更大乱斗_-赛博清理员v1.18/types.ts
export enum GamePhase {
  WELCOME = 'WELCOME',
  SETTINGS = 'SETTINGS', // New Phase
  DIFFICULTY_SELECT = 'DIFFICULTY_SELECT',
  CHARACTER_SELECT = 'CHARACTER_SELECT',
  STORY = 'STORY', 
  PLAYING = 'PLAYING',
  SHOP = 'SHOP',
  BOSS_INTRO = 'BOSS_INTRO',
  GAME_OVER = 'GAME_OVER',
  VICTORY = 'VICTORY',
}